PageWithProperties -> Page
